kerosene: Move getCalendarWeeks to new moment group
